mvneta: select phy by phy_select sysfs attribute
The phy_select is a writable string attribute. Writing the name of the PHY as
defined in the DTS will select that PHY for the next setup of the target iface.
The initial value of the phy_select is phy-def.

The phy_select accepts values prefixed by 'phy-' only to prevent using some
unwanted node. Eg.

  # ip link set dev eth1 down
  # echo -n phy-sfp > /sys/class/net/eth1/phy_select
  # ip link set dev eth1 up

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>