Omnia HACK: Omnia is shipped with weird ath9k card
The Omnia board is shipped with weird ath9k card that has 0x65
regulatory domain code in its EEPROM. It is simply wrong because it
limits TX power to 16 dBm and does not allow user to set US or EU
regulatory limits.

This patch allows that. It is hack. Better solution needed.